dev_requirements.txt
pytest-cov==1.6
dnspython==1.12.0
selenium==3.141.0
urllib3==1.10
requests==2.11.1
syncloud-lib==185
